combined the backend to be all one file so vercel can handle it properly
